+++ src/extension.ts
@@ -211,6 +211,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	subscriptions.push(vscode.workspace.onDidChangeTextDocument((TextDocumentChangeEvent) => {
 		const filePath = TextDocumentChangeEvent.document.uri.fsPath;
 		if (filePath && !filePath.toLowerCase().startsWith(`extension-output`)) {
+			MarkdownFoldingProvider.triggerHighlighting();
 			statusDebouncer(() => triggerShowDraftStatus(`onDidChangeTextEditorSelection`), 200);
 		}
 	}));