+++ src/helpers/ArticleHelper.ts
@@ -115,7 +115,7 @@ export class ArticleHelper {
     const lastLine = lines.pop();
     const endsWithNewLine = lastLine !== undefined && lastLine.trim() === "";
     
-    let newMarkdown = this.stringifyFrontMatter(article.content, Object.assign({}, article.data));
+    let newMarkdown = this.stringifyFrontMatter(article.content, Object.assign({}, article.data), document?.getText());
 
     // Logic to not include a new line at the end of the file
     if (!endsWithNewLine) {
@@ -150,8 +150,9 @@ export class ArticleHelper {
    * 
    * @param content 
    * @param data 
+   * @param originalContent 
    */
-  public static stringifyFrontMatter(content: string, data: any) {
+  public static stringifyFrontMatter(content: string, data: any, originalContent?: string) {
     const indentArray = Settings.get(SETTING_INDENT_ARRAY) as boolean;
     const commaSeparated = Settings.get<string[]>(SETTING_COMMA_SEPARATED_FIELDS);
 
@@ -165,7 +166,7 @@ export class ArticleHelper {
       }
     }
     
-    return FrontMatterParser.toFile(content, data, ({
+    return FrontMatterParser.toFile(content, data, originalContent, ({
       noArrayIndent: !indentArray,
       skipInvalid: true,
       noCompatMode: true,